Skip JUnit parsing when Ansible is ok
When qe-sap-deployment Ansible playbooks use block/rescue to handle
transient failures, the command returns exit code 0 if recovery is
successful. However, the JUnit reporter still records the initial task
failures in the generated XML reports.

Previously, parse_extra_log("XUnit", ...) was called unconditionally on
all generated reports. This caused openQA to mark the job as failed by
reading those recorded failures, even when the deployment was ultimately
successful.

This change restricts JUnit report parsing to cases where the Ansible
command actually fails (non-zero exit code), preventing false openQA
failures for successfully recovered deployments.

+++ tests/sles4sap/qesapdeployment/deploy.pm
@@ -103,9 +103,11 @@ sub run {
         '/tmp/results/',
         '-type', 'f',
         '-iname', "*.xml");
-    for my $log (split(/\n/, script_output($find_cmd))) {
-        parse_extra_log("XUnit", $log);
-        enter_cmd("rm $log");
+    if ($ret[0]) {
+        for my $log (split(/\n/, script_output($find_cmd))) {
+            parse_extra_log("XUnit", $log);
+            enter_cmd("rm $log");
+        }
     }
     qesap_ansible_softfail(logfile => $ret[1]);
     if ($ret[0]) {
